Bug#33181964 Incorrect result for pushed join w/ outer join inside an exists-subqry
When a pushed join contain index scan(s), we may need to return the result set in multiple batches.
In a join of the two tables 't1 JOIN t2', where we assume that t1 is a table scan, and t2
is joined with t1 using an index, multiple t2 rows may be returned for each t1 row.

In case all rows from t2 matching the supplied keys generated from the t1 rows cant be
delivered in a single batch, more t2 rows will be delivered in the next batch, which will
be matched against the _existing_ set of t1 rows by the SPJ API.

In an INNER JOIN between t1 and t2, the SPJ API will present this to the callee as if
the t1 rows raeppear with new t2 rows later in the same result set. This is OK as
the INNER JOIN operation itself does not implies any ordering of the rows in the
result set - As long as all t1 x t2 rows are returned we are OK.

For an OUTER JOIN and SEMI JOIN (Exist-subqry) however, the situation is more complicated:

- An OUTER JOIN may still return _matching_ rows in any order (over multiple batches).
  In addition it also need to keep track of t1 rows where no matching t2 rows where
  found and generate a NULL-extended 't1 | NULL' result rows for these when all
  matching t2 rows has been returned.

- A SEMI JOIN (aka: 'FirstMatch') is partly the opposite of the outer join: As soon
  as the first matching t2 row has been returned, any further matches should be ignored.
  Thus there will not be multiple occurences of the same t1 row as more t2 batches
  are retrieved. Furthermore, if no t2 matched were found at all, the t1 row should
  not be returned (Similar to an INNER JOIN)

It is the responsibility of the SPJ API to implement the above semantic for OUTER/SEMI joins.
(Note that the SPJ block may also know about the join type, but this just for optimization
purposes to avoid returning more rows than needed - Any 'extra' rows will be eliminated
by the SPJ API nevertheless.)

NdbResultStream::prepareResultSet() process the set of rows received from the SPJ blocks,
find any matching rows from the involved tables and applies the specified join types (INNER,
OUTER, SEMI or ANTI-join) to filter out rows _not_ to be returned.

In order to handle the OUTER- and SEMI-join semantic described above, the bitmask
m_matchingChild is maintained for each row in the results sets. Note the comments
for this member variable where it is defined in the TupleSet class:

............. Pasted from code, less relevant parts skipped  ....................

    /**
     * m_matchingChild keep track of current and previous matches found for
     * this tuple in the TupleSet:
     *
     * .... the bits in m_matchingChild keep track of match history
     * in previous result batches relating to this TupleSet. It has two usages,
     * depending on whether it is an outer- or firstMatch-semi-join:
     *
     * outer-join:
     *   The aggregated set of (outer joined) nests which matched this tuple.
     *   (NULL-extensions excluded.) <Only the bit representing the firstInner
     *   of the nest having a matching set of rows is set>. Needed in order to
     *   decide when/if a NULL extension of the rows on this outer joined
     *   nest should be emitted or not.
     *
     * firstMatch semi-join:
     *   The aggregated set of treeNodes which has a previous match with tuple.
     *   Used to decide if a firstMatch had already been found for this tuple,
     *   such that further matches should be skipped.
     *
     *  ...
..................................................................................

Note that the comment says that for outer-joins:

  'Only the bit representing the firstInner of the nest having a matching
   set of rows is set'

The failing query reported as part of this bug is __both__ a 'firstMatch'
and an outer join of two tables - in pseudo code:

  firstMatch (t1 outer join (t2 inner join t3))

The query is crafted such that there will be no matches from the t2 table,
note the on condition term: 'on t2.pk = 999', which will not match any
rows. Furthermore we can see from the the explain that both t2 and t3 are
'child of t1'.  -> We have the query tree: (OJ = Outer Join, IJ = Inner)

                    t1
                  /    \
            (OJ) /      \  (OJ)
                t2  IJ  t3

Using the same terms as in the NdbQueryOperation code:

  - (t2, t3) are in the same inner-nest (They are mutually inner joined)
  - t2 is the 'firstInner' in this nest.
  - The inner-nest of (t2,t3) 'isOuterJoin' relative to t1.

When ::prepareResultSet iterated the returned child rows for t1, it
will find that both t2 and t3 'isOuterJoin()'. If a t3 'childMatched'
(with t1) is found, the existing implementation will set the
m_matchingChild-bit corresponding to t3 in the t1-TupleSet. This
is the root cause of the bug, as the comment specifies that in order
to keep track of matching outer joins, only the bit corresponding to the
'firstInner' should be set -> That is: We do not have a match
of the (t2,t3) inner-nest until all tables upto t2 had a match.

Register t3 matches before t2 has been checked (which will never match),
incorrectly resulted in that we belived we had previously returned a
'firstmatch' when we had iterated all the t2,t3 row batches. Thus the
firstMatch logic skipped the entire t1 row, instead of falling through
to the outerJoin logic which should have generated a NULL-extended
[t1-row, (t2,t3)-NULL-extended] result row.

Fix is to check when we find a 'childMatched' being 'isOuterJoin', that
it also 'isFirstInner', before setting the m_matchingChild-bit. That
enforce the logic described in the comment.

Also note that we correctly _did_ check for 'isFirstInner' when we
decide whether a NULL-extended row should be created or not.

So failing to perform this check didn't affect the outer join logic, as
the incorrectly set bit was not checked at all in that part of the code.
However, it did interfer with the firstMatch logic which used these
other bits as well.

Reviewed by: Arnab Ray <arnab.r.ray@oracle.com>