Bug #33152269: No result set returned when using HAVING COUNT(DISTINCT ...)
Bug #31790217 (REWRITE STREAMING AGGREGATION) changed AggregateIterator
to be “slice-neutral”, ie., it no longer changed which slice was active
in the JOIN. However, this opened up a bug in an edge case: If there's a
sort before the aggregation, a temporary table after it, and a HAVING
condition (referring to aggregates), the HAVING condition could be
evaluated in the wrong slice, as neither the temporary table nor the
sorting would set up the correct one on Read(). This means that
join->sum_funcs (written to by AggregateIterator) would point to one set
of aggregation Items, but the field list (used in the condition by
FilterIterator, for HAVING) would point to a different one.
For aggregates with DISTINCT aggregators, this would actually matter,
and we would always get 0 from the condition (for COUNT(DISTINCT ...)),
leading to wrong results.

Fix by a hack: We force the AggregateIterator to refresh the slice on
each output. The hypergraph optimizer does not use slices, and thus is
not affected.

Change-Id: I75b6c4398fc6a836e30c62b6d956d2bbea1b7118