Bug #33123298 <CS>.LISTROUTERS() DISPLAYS INCORRECT ROUTER PORTS
When doing a bootstrap Router bind ports get written to the configuration file
and also registered in the metadata field routers.attributes.
However they are never updated in the metadata even when the user changes
the ports in the metadata.

This patch changes the behavior so that now all known router attributes are
updated when the Router starts the same way the version string was updated
so far.

RB: 26791