Bug #33132384 CLUSTERSET: ROUTER LOG SHOULD CONTAIN A MESSAGE INDICATING A CHANGE IN ROUTER OP
This patch adds additional INFO level log:

2021-07-29 09:34:58 metadata_cache INFO [7ff5ac460700] New target cluster
read from the metadata: 'main_cluster'

It will appear in the logs only if the target cluster for a Router is
actually changing.
So for example: if cluster A is a PRIMARY, our target_cluster is "primary" and
someone changes the target cluster from "primary" to A the Router will not
log it as effectively the target cluster did not change for the Router.
On the other hand if the target_cluster is set to "primary" and there is
a Primary Cluster change in the ClusterSet the message will be logged as there
is effective change in the Router's target_cluster even though no
setRoutingOperation has been performed.

RB: 26776