Bug#33193304: MAIN.RANGE_WITH_MEMORY_LIMIT fails sporadically ON PB2
Test file range_with_memory_limit.test was added as part of Bug#17769777.
It has one test case related to the bug and then runs the entire set of
tests for range optimizer using a low memory limit. This makes some
of the queries to not pick range scan because the memory limit would
exceed. However, this is not deterministic because the memory limit
could suffice to pick range scan on certain platforms and on some other
it might not be. So the test has never been stable.
Also, the original intention of the test to run all queries in range.inc
was to have more test coverage since the memory checks were added in all
parts of range optimizer. However, with the consistent complaints of
not being stable on all platforms, the memory limit is currently set to
so low that the first allocation in range optimizer itself fails most of
the time, thereby making most queries not to pick range scan or even try
to pick range scan.
So, we have removed the test file and moved the one test added for
Bug#17769777 to range.inc.

Change-Id: Id23fafe2ecf502c7c6de163f17a3ea1937ec35b2