Bug#33141979: Field::val_time_temporal() const: Assertion `0' failed
A YEAR field is attempted to be interpreted as a temporal value,
however there is no implicit conversion between those two.
The current complication is that a YEAR value has is_temporal()
returning true (YEAR is a temporal data type), however it is not
represented internally with a MYSQL_TIME value.

Fixed by checking that a temporal value is not of data type
MYSQL_TYPE_YEAR. It is a bit of a kludge, but the situation will become
better when we use data_type() as the sole determinator for how to
perform comparisons.

Reviewed by: Catalin Besleaga <catalin.besleaga@oracle.com>