Use system CA roots for client if no option provided
Right now a CA chain is required to be provided in verify_ca or
verify_identity mode. This makes it much more cumbersome when writing a
client that connects to a server where the certificate in the system
roots and when it provides a proper certificate signed by such an
authority.

By falling back to the system roots, deployments are greatly simplified
because it removes the need to manually specify the root chain. The root
chain is highly operating system and deployment dependent, so it
requires a different configuration flag on each platform.

By setting up the system roots through OpenSSL, deploying a MySQL server
with a root signed installed on a system is greatly simplified for
client configuration.

Many clients in other languages / on other platforms already follow a
similar behavior where system roots are used if no specific CA chain is
configured. This applies to for example Java, Go, Node.js, Rust & .NET.

This change also makes it simpler for drivers that use libmysqlclient
such as the ones often used in Ruby and Python to simplify configuration
for developers there as well.