BUG#31716706 BAD HANDLING OF QUOTES IN SHOW GRANTS
Description: SHOW GRANTS quotes username with ' and if the username
contains ' it is not escaped. When the output of SHOW GRANTS is
dumped to a file to re-excecute the commands will result in either
invalid SQL or SQL injection.

Analysis: It is hardcoded to quote the username with ' and also fails
to escaping the ' when present in username. Right way is to append the
quotes and escape it with \ when quotes are present in username or host.

Fix: Escape the ' with \ if present in username@host specification.

RB: 26845