Bug#32851684: Segmentation fault when using VALUES
The problem occurs when attempting to use a VALUES clause as a
scalar subquery. When evaluating an Item_values_column, the m_value_ref
member is not initialized, thus the program fails.

However, using a VALUES clause as a scalar subquery is only valid
when the VALUES clause contains a single row. In this case, the members
of the VALUES are substituted into the outer query, thus eliminating
the Item_values_column object. We can also safely reject VALUES clauses
with more than 1 row already in resolving.

Change-Id: I3fcbba6e97752f90d404d0c46c2b678a056a0764