Bug# 31117893 - GROUP BY WILL THROW TABLE IS FULL WHEN TEMPTABLE MEMORY ALLOCATION EXCEED LIMIT
In the case of a query that uses temporary table for aggregation,
if the group-by field value is already present, the row is updated.
Otherwise, a new row is inserted into the temporary table.

If the temporary table has variable length fields, like varchar,
the TempTable implementation may exhaust the memory for an update
operation.

This out-of-memory error from the temporary storage engine is not
handled during an update operation and causes the error.

The fix is to move the temporary table to disk and retry the update
operation when a out-of-memory is reported.

Change-Id: I06ab81923150262158c7eb3fe1353a4ff573da0b