Bug#32345941 - SELECT RESULT IS MISMATCH WHEN USE COALESCE AND NOT.
There is an inconsistency in how certain comparison predicates are
evaluated, e.g. when part of a WHERE clause. Assume we have a query:

SELECT ... FROM <table> WHERE <decimal column>=<string expression>

where:
* We use an index on <decimal column>.
* <string expression> evaluates to a prefix p followed by a suffix s.
* p is a valid decimal literal.
* s cannot be parsed as part of that literal
  (e.g. <string expression> is '1.3hello').

Then if <string expression> is a string literal, the where clause
evaluates to 'true' whenever <decimal column> is equal to the decimal
value of p.  But if <string expression> is a function
(e.g. CONCAT('1.3', 'hello')), the WHERE-clause is always false.

Likewise, if <string expression> is an empty string, it evaluates to 0
if <string expression> is a literal (i.e. ''), but if it as a
function, the WHERE-clause is always false.

This commit fixes that inconsistency, by ensuring that
Field_new_decimal::store() does not flag E_DEC_BAD_NUM as an error.

This commit also:
- Adds test cases.
- Makes a number of identical test queries in main.mysqltest visibly
  different, so that it is easier to debug failing test cases among these.

Change-Id: I2c1e51c31fd44afcc524a33750d31b9c846c5e95