Bug #33037007: PRELIMINARY FIXES FOR WL #14488 [min_functions, noclose]
For TRP_GROUP_MIN_MAX, extract {min,max}_functions when we create
the TRP, instead of when we create the QUICK_SELECT_I. The reason
is that the latter makes the extraction strongly dependent on
when in the query optimization phase the object is created,
as join->sum_funcs changes once temporary tables have been set up.
Thus, making it more independent is a prerequisite for moving the
QUICK_SELECT_I creation to CreateIterator() time.

Change-Id: Ib3d3c0f4f63b47434962dc4623df1fdc390013e7