Bug#33079073 - SQL_CMD_CALL::EXECUTE_INNER(THD*): ASSERTION `THD->IS_ERROR() || THD->KILLED' FAILED
Assertion to check if error state is set in the diagnostics area
on ANALYZE TABLE to UPDATE HISTOGRAM non-success in SP is
failed in the debug build. In the non-debug build, statement
executes successfully.

ANALYZE TABLE sends errors stored in the diagnostics area
as a result set, clears diagnostics area and returns success
return value to the caller. But ANALYZE TABLE TO UPDATE
HISTOGRAM incorrectly returned non-success return value after
clearing the diagnostics area. Hence assert condition to check
if error is set in diagnostics area failed.

Updated code to return success value in this case and errors
in the diagnostics area are still sent as a result set.

Change-Id: I219e76fd8376627164f3dbf3dea5594b5c500cf7