Bug#33197276: regression: Assertion `cur_query_block != select' failed.               and segfault in Condition_pushdown::replace_columns_in_cond Bug#33209907: condition seems not correct
While pushing a condition down to a derived table, we clone a view
reference. To resolve this cloned view reference, we use the merged
derived table's context when a Item_view_ref is created for merged view
fields. If not, which is the case for a natural join column, we use the
context of the identifier itself. However, if this view reference is used
in a subquery having no tables which gets merged later, then the context
of the identifier is not updated correctly to be using the parent's query
block context. Instead, while merging contexts (when the subquery gets
merged), the context's query block would be updated. However, the tables
to be used to resolve the field would still be that of the subquery's
which in this case are none.
As a result, during resolving of the cloned view reference, it tries to
resolve the field as an outer reference, but fails because for an outer
reference, the query block of the context should be different from that
of the query block in the outer context which in this case are same.

We now update the identifier's context in fix_after_pullout() to that
of the parent's query block if that identifier was resolved in the
parent query block.

While at it, fix Bug#33209907 by checking for the correct return value.

Change-Id: Ic4b80cd966a4da5898ab5bb60f3178cedda551fe