Bug #33037007: PRELIMINARY FIXES FOR WL #14488 [record, noclose]
QUICK_SELECT_I has a public member uchar *record, that is used in place
of m_table->record[0]. This gives the impression that it is mutable
during execution (even from external callers), and indeed, ROR scans
do modify this buffer with a comment saying that all ROR scans must have
the same buffer. However, this is a red herring; the pointer is always
the same, as all ROR scans share the same TABLE object (although not the
same handler), and thus, the same record[0].

Replace with m_table->record[0] everywhere, and remove the member.
It's highly unlikely it has any kind of performance benefit.
Same with m_table->record[1].

Change-Id: Icd103e3859e637fdf2a8e3d3a7b5b860efaab00c