Bug#33253367 Wrong build dependency for checking autotest testcase files
Building with ndbcluster using cmake 3.20 warns about policy CMP0115.

    CMake Warning (dev):
      Policy CMP0115 is not set: Source file extensions must be explicit.  Run
      "cmake --help-policy CMP0115" for policy details.  Use the cmake_policy
      command to set the policy and suppress this warning.

      File:

        ${CMAKE_SOURCE_DIR}/storage/ndb/test/run-test/atrt.hpp
    This warning is for project developers.  Use -Wno-dev to suppress it.

The MAIN_DEPENDENCY option to cmake ADD_CUSTOM_COMMAND command is
strictly adding a source file dependency not a target dependency.

MAIN_DEPENDENCY atrt was interpreted as a dependency on atrt.hpp not on
the target executable atrt as intended.

Solution is to move atrt dependency to the DEPENDS option which allows
also target dependencies.

Change-Id: Ice3d6914f478aa87243bb798ce76bfd1cff89b65