Bug #33264498 /permissive- and Implicit conversion of integral constant expressions to null po
The clang on windows build is broken. The underlying problem is the
missing implementation of "/permissive-" in clang.

See the corresponding bug filed for LLVM:
https://bugs.llvm.org/show_bug.cgi?id=51582

The workaround is to remove 'constexpr' from the definition of
PSI_NOT_INSTRUMENTED.

Also: change cmake code, removing compiler options that are
valid for Visual Studio only.

Change-Id: I83f6b6605b14722d7d58f0076afee7a96c3bd68d