Bug#33262452 Running more that 64 node cluster can crashes in Windows
The set_max_count() is an Utility function for dynamically expanding
the fd_set on Windows to get around the hardcoded FD_SETSIZE limit.
But, there are few problems with the function.

1) The first parameter, set should be an input/output parameter. The
memory allocated by the set_max_count() will be used by other function
in the header. But, the set is used as a single pointer. So, the
reference to the memory allocated will be lost once we move out of the
funciton. Fixed it with a double pointer

2) The expression in the malloc misses a parenthesis, because of this
the memory allocated will not be sufficient and any attempt to set
socket to fd_set will lead to memory corruption. Fixed by adding
proper parenthesis.

Change-Id: I72577f3ca24a8ade52ee2c2069025db5796e7a90