Bug #33264928: Multi-equality is not fully applied by hypergraph optimizer with cycles
If a multiple equality involves several fields from one table
and at least one field from another (but not three or more tables),
the hypergraph optimizer could expand it so that the join conditions
on a hypergraph edge would be the only thing responsible for
enforcing the intra-table equalities. If that join edge would then
be elided due to not being chosen as part of the spanning tree of
the hypergraph (ie., there is a hypergraph cycle, and we happened
not to choose that edge due to a lower cost somewhere else), we could
lose that equality.

We fix this by prioritizing expanding intra-table equalities when doing
early expansion of multiple equalities. We would already do this if
the multiple equalities involved a constant, one table or three or more
tables, but not if they were exactly two, and the ordering of the
tables within the multi-equality was unlucky. This should also make it
easier to push such conditions down into table filters, which is usually
(but not always) what we want.

Change-Id: Ibb7f9db2d1113d6f0c1328a7b9716ac1dd36c02e