Bug #33212432 thread stack overrun/segfault on ASAN builds, Item_func_minus::int_op with EXTRACTVALUE
In check_stack_overrun(), ensure that we have code which uses the
'buf' argument. Otherwise the stack variables in the calling functions
may be optimized away.

No test case, as it is very hard to write one that is deterministic on
all platforms, and all kinds of builds/optimization levels.

Change-Id: If61343d9eae0c67c3b9c94fc31278dc121775ad6