BUG#33270401: Uninstall not blocked for active old-named semi-sync replication plugins
* Background

MySQL 8.0.26 (WL#14194) renamed much of the replication terminology,
to use source/replica instead of master/slave.

To keep backward compatibility for the semi-synchronous replication
plugins, we kept the existing plugin libraries using old terminology
in plugin names, filenames, and other identifiers, while introducing
new plugins with new terminology both the plugin names, filenames, and
other identifiers.

* Problem

The rename broke the existing code that blocks the uninstallation of
the old-named plugins, because it was coded in a way that depended on
the plugin names.  That code had been auto-translated to check for the
new-named plugins, so the check stopped taking old-named plugins into
account.

* Analysis

The underlying problem was that the semisync plugins used an
inadequate method to check if uninstallation is allowed. The check was
hard-coded in the generic plugin framework, in sql_plugin.cc, checking
if the plugin name matched the semisync plugins.

* Fix

Define the existing plugin hook, 'check_uninstall', for the semisync
plugins.  This will be invoked regardless of the plugin name.  The
code for this is shared betweeen the old-named and new-named plugins,
so now the check is executed regardless of plugin names.

RB:26885