BUG#33267294 - clusterset: router target_cluster global/default option does not exist
This patch removes the logic where the Router tries to determine what should be
the target cluster if neither --conf-target-cluster nor
--conf-target-cluster-by-name are provided on bootstrap. It is now up to the
metadata configuration (clustersets.router_options.target_cluster) what
should be the default target cluster in that case.

This patch also makes use of the v2_router_options view to query for router
options (like target_cluster).
This view hides the fallback to the clustersets.router_options in case of
option missing in routers table.

RB: 26893