Bug#33132348 CLUSTERSET: REBOOTSTRAP NOT NEEDED WHEN CREATING CLUSTERSET, CONFIG INCORRECT
This parch changes the behavior when the Router was bootstrap was bootstrapped
against a standalone Cluster and it sees in the runtime that now the Cluster
is a part of a ClusterSet.

Now it logs a warning suggesting re-bootstrap and keeps using the original
Cluster as it was still a standalone Cluster.

In order to know what was the original Cluster type during the bootstrap
now the Router stores tat information in routers.attributes.'$.bootstrapTargetType'
field.

RB: 26912