Bug#33177686: mysqld_list_fields leaks memory! Bug#33013097: com_process_info fails
The problem in 33177686 is that mysql_list_fields() is using a
non-standard way to execute an SQL construct and misses a cleanup of
temporary tables created to evaluate JSON table functions underlying
a view.

Fixed by calling LEX::destroy() after the SQL construct has been built
and used for listing fields.

The patch also fixes bug#33013097 which is due to a dangling
query_result pointer after listing fields.

Reviewed by: Dyre Tjeldvoll <Dyre.Tjeldvoll@oracle.com>

Change-Id: I6d2c1c82ff877992c73c73a7fd5e3787843325e2