Bug#28658625 NDBINFO QUERY USED BY MEM PERFORMS BADLY 2/3
Backport of this fix to Cluster 7.5.  Only the first two
of the three patches are backported.

Step 2: provide size estimates for all ndbinfo tables.

Add a rows_estimate column to ndbinfo.ndb$tables. In Ndbinfo.hpp,
add a std::function element to struct Ndbinfo::Table::Members, and
in NdbinfoTables.cpp, add to each table declaration a function that
will calcuate an estimated number of rows for the table.

Each estimate is put into the rows_estimate column of ndb$tables,
where the mysql server uses it to populate table statistics for the
optimizer.

Change-Id: I8e1e9502ea6f04cce76ff65440ae7e1518d6e542