Bug#33002479 "HIGH MEMORY ALLOCATION FOR DROP DATABASE" [noclose]
Attempt to drop database with many tables with thousands of partitions
consumed high amounts of memory and could have resulted in out-of-memory
condition.

With transition to New Data-Dictionary and atomic DDL, DROP DATABASE
started to allocate much more memory which is released at the end of
statement. This memory allocated for various DD API objects, new MDL
locks, temporary objects used during updates to DD tables and so on.

Part of them problem was that the code responsible for acquiring
metadata locks on the tablespaces of tables/partitions to be affected
by DROP DATABASE, allocated MDL_request objects on statement's runtime
memory root (which was not freed until statement end), even though
these objects were not used after the locking phase.

This patch solves the problem by introducing special memory root, which
is used for allocating MDL_request objects for such tablespaces, and
which is destroyed right after locks have been acquired.

There is no MTR test case as it is hard to write robust one for this
situation.

This is one more step towards solving the original big problem to the
full extent.

Reviewed-by: Dyre Tjeldvoll <Dyre.Tjeldoll@oracle.com>