Bug #33037007: PRELIMINARY FIXES FOR WL #14488 [convert range scans, noclose]
Convert the range scan iterators to have native Read(), which removes
the entire need for the QUICK_SELECT_I::Read() wrapper, and get_next().
QUICK_SELECT_I is now a pretty meaningless empty shell.

Note that this changes the definition of rows_examined for ROR scans;
we now count all rows that have been returned in the base scans,
not the number of rows that come out of the union/intersection operator.
This seems like a more reasonable definition of an ”examined” row
(similar to how we count rows that have been removed by a filter).

Change-Id: I2e84db9be48a7b7313a2db716314fbc6a2e95011