Bug#33064062 CREATE...SELECT STATEMENTS BREAK REPLICATION WHEN ANSI MODE IS ENABLED
CREATE TABLE SELECT which created InnoDB table was incorrectly written to
binary log if binlog_format was set to ROW and sql_mode contained ANSI. As
result replication of such statement failed with error on replica.
Application of such binary log using mysqlbinlog tool failed as well.

The atomic CREATE..SELECT was implemented by adding a new clause to create
table called START TRANSACTION. The problem was that this clause was NOT
added when ANSI sql_mode is enabled.  As a result, the applier tried to
execute ordinary implicit-committing CREATE TABLE in the middle of
transaction. And the latter causes an error in GTID mode where transaction
has an assigned GTID.

There is no good reason why addition of this new clause should depend on
SQL-mode. So this patch solves the problem by removing such dependency

Change-Id: I31594e7a08c5b807046a10288fd42351b5df3b7d