BUG#33276021 INNODB PARALLEL SCAN reports incorrect partition ID
Problem :
---------
While initializing parallel scan we are trying to get the partition ID
from DD partition table partition number. This works if there is no
sub-partitions. For sub-partitioned table, this is the sub-partition
sequence number within a partition and not the partition ID.

Solution :
----------
The partitioned table is already open at this point and the partitions
are organized based on the partition ID sequence number. No need to
access DD. The index is the partition ID.

Reviewed-by: Sunny Bains <sunny.bains@oracle.com>

RB: 26907