Bug #33037007: PRELIMINARY FIXES FOR WL #14488 [internal AccessPath, noclose]
Use AccessPath more places internal to the range optimizer;
i.e., all the submodules that used to return TRPs now return
AccessPaths. This is a step on the road to splitting out individual
return types. Like the previous patch, there are several places
where we break the AccessPath encapsulation, that will need to be fixed.

Note that this fixes a cost model bug that has been present
since at least 2004; roru_intersect_part was multiplied by
the result of an int/int division, that was nearly always zero.
Now that we use AccessPath's num_output_rows (which is double)
instead of the TRP's records field (which is ha_rows),
this bug goes away, and a few EXPLAIN results change.

Change-Id: I11277db6674b6d23244671be968bd48ef34100c6