Bug#33275457: Fix multi-valued index
Patch #3: Multi-value indexes are not used for MEMBER OF combined with OR.

If we have a MEMBER OF predicate that is combined with other predicates
with an OR operator, the range optimizer will not consider MEMBER OF
as a candidate for using an index.

Fixed by adding a case construct for Item_func::MEMBER_OF_FUNC in
get_func_mm_tree().

json_overlaps() and json_contains() already has this type of support.

This is a contribution by Yubao Liu.

Change-Id: I1fd7a78091998437310973b3c24099ad554a58a6