Bug#33275457: Fix multi-valued index
Patch #4: Wrong results with AND of JSON_CONTAINS and multi-valued indexes

For a multi-valued index on json array field f=[1, 2], this statement
wrongly returned empty result set:

  SELECT * FROM f WHERE JSON_CONTAINS(f, 1) AND JSON_CONTAINS(f, 2);

The cause is that get_func_mm_tree() converts two JSON_CONTAINS() to
`f_idx = 1 AND f_idx = 2` which is always false for single value index
but possible for multi-valued index.

Fixed by anding the two key ranges, rather than marking the condition
as always false.

This is a contribution by Yubao Liu.

Change-Id: I535fc6ce8755f4f3b6e8cbd77b4c0ee4aa685cae