Bug #33037007: PRELIMINARY FIXES FOR WL #14488 [retrieve_full_rows, noclose]
Of all the range scan types, ROR intersection is the only one
that actually honors retrieve_full_rows (all others either always
or never do index-only scans); it is set to false when the intersection
is part of a ROR union. Thus, make it a parameter specific to
that one class instead of a generic parameter in make_quick().

Change-Id: I2f29ae14566bb3092e2e9794a65ce36ecbfadbaf