Bug#33276418: A node reports itself as RECOVERING even after it fails to auto-rejoin the group
Description: During auto-rejoin member state is set to RECOVERING.
If view-change takes time a member leaves the group.
However, the member state is not being changed back to ERROR causing the member
state to be shown as a state set by view-change.

Resolution: If view change takes time, group replication leaves the group.
So irrespective of the state of old view change, member state should be set to
ERROR.

ReviewBoard: 26928