Bug#33087212: COMPONENT COMPONENT_TEST_BACKUP_LOCK_SERVICE CAUSES CRASH AT SHUTDOWN.
Problem: Issuing a shutdown while the component_test_backup_lock_service
was loaded resulted in SEGV.

Root cause was that the unload hook for the test component did not
handle being called during shutdown with current_thd==nullptr, (all
components are unloaded when the component system is stopped as part of
shutdown).

Solution: Avoid dereferencing the pointer provided by current_thd if it
is nullptr.

Change-Id: Idc008f1ea8f77488fa4f8c6e30329ec65bc574e6