BUG#32244822 CRASH IN TRACE_EXEC.ADD_SELECT_NUMBER(UNIT->FIRST_QUERY_BLOCK()->SELECT_NUMBER)
A DML statement crashes on accessing a subquery which has been
optimised away and cleaned up. The DML statement has a COUNT which
contains a subquery on a const table. The COUNT is removed and
replaced with a constant, and then cleanup is triggered. This also
cleans up the argument of the COUNT, which is the subquery. After
the cleanup, the list of sum-functions is built and an aggregate is
added for the COUNT based on the field list. While generating access
paths, the aggregate's argument subquery is accessed, resulting in a
crash since the subquery no longer has a valid query block.

The fix is to replace COUNTs with hidden NULLs in the field list of
the aggregate query block. This is done when the COUNT is being
cleaned up after its removal.

Change-Id: I78ca635c32c90793cb033c08563813d8d1add038