Bug#32873315 - IMPROVE LOGGING IN XCOM AROUND JOINING NODES
Description:
-----------
XCom has a very sensitive area around joining nodes and internal node
recovery, which is the fact that if the nodes that are in the group
are not able to contact the joining node or the joining node does not
receive the whole information, it might stay in a weird limbo on the existing
group.

As a first step, we need to improve logging around those recovery operations.

Suggested Fix:
-------------
Improve logging around a node joining a group.

RB: 26950