Bug#33197418 Assertion `((*border_ptr)->fixed && m_frame->m_query_expression == WFU_RANGE) || ((!(*border_ptr)->fixed || (*border_ptr)->basic_const_item()) && m_frame->m_query_expression == WFU_ROWS)' failed.
A constant subquery in RANGE INTERVAL expr was not handled correctly:
during fix_fields in Window::setup_range_expressions, the expr was
changed from Item_singlerow_subselect to Item_int.  This was not
reflected in the alias PT_border::m_value, which caused crash or
assert in Window::check_constant_bound.

Fixed by updating it if changed by fix_fields/

Change-Id: I263996457c0930b9cbbeb05d56cb138ea0a9c603