Bug#33250020: regression: FullTextSearchIterator::Init(): Assertion `m_ft_func->ft_handler != nullptr' failed.
When performing a full-text search query against a table stored in the
blackhole storage engine, and the chosen query plan used a full-text
index scan, an error was raised in release builds, and an assert
failure was raised in debug builds.

The expected behaviour is that the query returns an empty result. This
is what happens if the chosen query plan uses a table scan.

Fixed by adding more stubs to ha_blackhole; just enough to allow
initializing, reading from and closing a full-text index scan.

The patch also cleans up some error handling in the full-text search
code:

- FullTextSearchIterator::Init() ignored errors returned from
ft_init(). Not any more.

- Item_func_match::init_search() did not stop execution if
ft_init_ext_with_hints() returned nullptr. This is why an assert
failure was raised in debug builds. Now, it stops immediately when
this happens.

ha_blackhole used to return an error from ft_init() and nullptr from
ft_init_ext_with_hints() before the fix, and this went undetected
because of the missing error handling. Now, ha_blackhole does not
return error/nullptr from those two functions, but the error detection
code is good to have in place even if it's not strictly required for
fixing this bug.

Change-Id: I233a35f7be156c876b6eaff7b66c81430bff51e3