Bug#32427727: Functional index ignored by SELECT query from within stored function
The problem was that substitution of expressions with generated columns
was effective for comparison predicates where the other operand was
a constant value only. In WL#9384, arguments passed to a procedure
was no longer regarded as a const value, but was denoted
const-for-execution, meaning that the value might change between
executions. However, the code for GC substitutions was not changed
accordingly.

The problem is fixed by replacing all relevant invocations of
Item::const_item() with Item::const_for_execution().

Original test cases have been complemented with similar tests that
substitute constant strings with user variables, showing that we see
the same plans.

Change-Id: I80f2d1985a7ab08b17ac3818bb5a8a1cbdf2985c