BUG#33317320 Lock contention index stats
When opening NDB tables for query, the index statistics are retrieved to
help the optimizer select optimal query plan. Each client accessing the
stats will acquire the global index stat mutex both before and after
accessing the stats. This causes mutex contention affecting query
performance regardless of whether queries are operating on same or
different tables.

Fix by protecting the index stat reference counting with an atomic
counter.

The problem was clearly visible when benchmarking with more than 32
(sysbench) clients, the throughput does not increase with increased
number of clients. With this patch the throughput continues to scale
also with 64 clients.

Change-Id: Iaf0f9afb096bd5caf851aab1cda67021e813a476