Bug#33079073 - SQL_CMD_CALL::EXECUTE_INNER(THD*): ASSERTION `THD->IS_ERROR() || THD->KILLED' FAILED
Followup patch to fix rpl.rpl_histograms test failure on pb2.

ANALYZE TABLE stores errors as a result set and returns success
value. But ANALYZE TABLE TO UPDATE HISTOGRAM, incorrectly returned
non-success return value in case of error. With the original patch
for this bug, ANALYZE TABLE TO UPDATE HISTOGRAM returns success
return value.
In error cases, transaction is rolled back and ANALYZE TABLE is
written to binlog. Other ANALYZE TABLE operations already
have checks not to release GTID ownership in this case. Such checks
are missing for UPDATE HISTOGRAM, and GTID ownership is released.
GTID ownership must be held when statement is being written to the
binlog. Writing to binlog without GTID ownership resulted in the
assertion failure. To fix this issue, GTID ownership is not released
on error even for UPDATE HISTOGRAM.

Change-Id: Icbec5a909d14541468e52e2c34663b4489d2d990