Bug#33250020: regression: FullTextSearchIterator::Init(): Assertion `m_ft_func->ft_handler != nullptr' failed.
Post-push fix for test failures in rapid.explain_dict and
rapid.explain_varlen.

The original patch changed the only caller of the
handler::ft_init_ext_with_hints() function to assume that an error had
been raised if it returned nullptr. This assumption holds for the
storage engines that implement full-text search.

However, the default implementation, used by storage engines that do
not support it, returns nullptr without raising an error. This is
usually not a problem, because there are resolve-time checks that
reject use of full-text search with storage engines that don't support
it, so the default implementation is not expected to be called.

But it can happen if the table has a secondary storage engine.
Specifically, if the preparation and resolving is done against the
table in the primary engine, which supports full-text search, and the
execution is done by the secondary engine, which does not support
full-text search, the default implementation gets called. And since
the code now takes an error path, even though no error has been
raised, it ends up in a confusing state which results in assert
failures.

Fixed by making the default implementation of ft_init_ext_with_hints()
raise an error.

Change-Id: I2154ea21908b22a8fcee7db9deca1655be734d62