Modify 'The table '...' is full message to indicate the configuration setting.
This error message in 8.0 has caused quite a lot of confusion and to the
developer it is not clear why the table is full or what they can do about it.

While the developer can not probably adjust the setting directly providing
the related configuration setting provides a pointer to the limitations
and settings that can be modified if needed.