Bug #33330728: HYPERGRAPH: ASSERTION `RIGHT_PATH->TYPE == ACCESSPATH::ZERO_ROWS' FAILED
We assume that if there's an impossible join condition, it cannot be
on an inner join; otherwise, the join condition would have been
pulled up higher and eventually converted to a WHERE (rendering the
entire query into an impossible zero-row query). An assert effectively
depends on this; it assumes that the right path has to be converted
to ZERO_ROWS, not the left path (ie., they are never switched around
due to commutativity).

However, due to a pass ordering problem before the join optimizer,
this isn't always true: Since we convert outer joins to inner joins
after pullup (ie., inverse of pushdown), we could have an outer join
converted to inner join, but the impossible join condition would still
be left on that join. Thus, relax the assertion.

Change-Id: Ifd4ab610c5d02173085d657397f8ba55cf63fecd