Bug #31894023 THE RESULTS OF AGGREGATION ON BIT TYPE ARE INCONSISTENT WHEN JOIN TYPE DIFFERS
Contributed by Hope Lee

A DML statement which uses an aggregation on a BIT type caches the
BIT values. BIT types are cached using an integer type, and later
looked up and converted to a string format to output the value. The
lookup treats the BIT value as an integer, which results in an
incorrect string value.

This is fixed by adding a new class for cached BIT values which can
convert bit values to string formats correctly.

Change-Id: I8394a6f395befae8626252596c025d422c81faa1