Bug#29268656: BOGUS CONSTANT, RANDOM OR TIMEZONE-DEPENDENT               EXPRESSIONS IN (SUB)PARTITIONING FU
Description
===========
Creating a table with non-deterministic functions in the
generated column expression, should not be allowed as per
the documentation, but this was not being enforced.

Additionally, creating such a table with partitions whose
expression did not actually refer to the problematic column,
failed with a misleading error message about the partition
expression.

Analysis
========
During create table statement parsing, if there are any
unsafe expressions defined for a generated column then
thd->safe_to_cache_query will be set to false. This variable
is later checked while parsing the partition expression.
This check will report an error even if the partition
expression does not refer to the problematic generated
column expression.

Fix
===
Reset the thd->safe_to_cache_query value before parsing the
partition function.

Change-Id: I1a45f2f06b73432fad12eb64357fe4edbb6c0ed1