BUG#33236909 Lock_time computed differently by slow log and performance_schema
Problem
=======

In the performance schema statement instrumentation,
the column LOCK_TIME in table performance_schema.events_statements_current
reports a value which is under evaluated.

By comparison, the slow query log also reports
the time spent waiting for locks,
and reports a bigger value for the same statements.

Root cause
==========

The LOCK_TIME is represented internally in the SQL layer
with THD::utime_after_lock.

The performance schema statement instrumentation is
updated only the first time member THD::utime_after_lock is set.

There is code in the server, however, that modifies
THD::utime_after_lock multiple times during the statement execution,
and these modifications are not reported to the performance_schema.

As a result, the LOCK_TIME reported in the slow query log
is greater than the LOCK_TIME reported in the performance_schema,
causing the bug.

Fix
===

The maintenance of THD::utime_after_lock was very convoluted,
creating code difficult to maintain and understand.

Instead of maintaining a timestamp, THD now maintains a delay
representing the (cumulative) time spent waiting on locks,
in member THD::m_lock_usec.

All the code has been adjusted to use THD::m_lock_usec instead.

When m_lock_usec is changed, the performance schema statement
instrumentation is updated as well.

THD::utime_after_lock was also artificially updated then restored
when printing a record in the slow log, in case of log throttle.

The code has been refactored to pass clearly parameters to
the slow log printing function, as a cleanup.

In addition, the LOCK_TIME previously only accounted for the time
spent waiting on the first set of table found, when locking tables.
This ignored subsequent calls to lock table,
hiding LOCK_TIME spent on log tables or in the data dictionary.

With this fix, LOCK_TIME accounts for:
- all the time waited on SQL TABLES
- all the time waited on DATA locks
and is reported consistently in the slow log and the performance schema.

Change-Id: I6764bbec4da88eeca053fc7724fdb4fb357badb7