Bug#33343948: Multi-valued index ranges are printed as binary
EXPLAIN FORMAT=TREE hex-encodes ranges for multi-valued indexes, even
if the data type is not a binary one.

The reason is that print_key_value() looks at the Field object
representing the array, whose type is an array type, whereas the key
values are actually of non-array types.

Fixed the printing by unwrapping Field_typed_array objects and using
the so-called conversion field which represents the elements of the
array.

A related problem fixed by this patch is that ranges using string
types were printed with hex-encoded values too, if they had a binary
collation. Whereas this also affected regular indexes, it was more
visible with multi-valued indexes, because multi-valued string indexes
always have utf8mb4_0900_bin collation. Changed so that hex-encoding
is used only for string types with a binary character set. Strings
with non-binary character sets are printed as plain text, but special
characters are escaped.

Change-Id: I2ebaf2cdc2060d510c358d7f4535dee9d6b11669