Bug#28658625 NDBINFO QUERY USED BY MEM PERFORMS BADLY 1/3
Backport of this fix to Cluster 7.5.  Only the first two
of the three patches are backported.

This set of three patches addresses ndbinfo JOIN performance,
first by providing more information to the MySQL optimizer via
the handler interface, and then by adding a primary key to some
on the ndbinfo base tables. This is a back-port to Cluster 7.5
of the original patch based on MySQL 8.0.

Step 1: provide row counts for virtual tables

A small set of tables is built in to the ndbinfo engine at
compile-time. The exact row count of these tables is always
known. There are three functional changes made here:

 - For the compiled-in tables, store the row count in
   stats.records, and set the HA_STATS_RECORDS_IS_EXACT flag.
 - Set stats.table_in_mem_estimate to 100% for the virtual tables
   and to 0% for other ndbinfo tables.
 - Calculate an estimated stats.mean_rec_length for all ndbinfo
   tables.

A new test case ndb.ndbinfo_plans is introduced here, and will evolve
through this series of patches to show the improvement to the query
plans.

Change-Id: Iebfddde34b57e4a58b772d4a47e04029f4bb1bf5