Bug #33159353: SHOW CREATE DATABASE fails with a role that has global level                privilege
Problem: A connected user which has an active role with a global level SELECT
privilege is denied access to mysql database.

Fix: When role is assigned to user with required privilege, we should also check
for Security_context::m_master_access, similar to how it is checked for user
account granted with required privileges.

RB#26922