Bug #33358527: Hypergraph optimizer double-counts joins from multiple equalities
In hypergraphs with cycles, we will have two sets of join conditions:
Those that come from the edge we are joining against (call them,
for the sake of the argument, primary edges), and then join conditions
from other edges that connect the two partitions (call them supplemental
edges). Some of these conditions may already be pushed down to indexes
(push to sargable).

We already had code to deal with a situation that could arise here,
namely that our primary edge was redundant with regards to other,
sargable-pushed predicates (due to coming from the same multiple
equality). We would refuse to join those specific paths along such an
edge (because it would be redundant). We would also have code to
skipping filters for sargable-pushed conditions on primary edges in
nested loop joins.

However, there was another situation we did not properly deal with,
namely that a _supplemental_ edge could contain a join predicate that
was already applied through sargable push. This would cause such filters
to be applied redundantly, and their selectivities double-counted.
Extend the logic of skipping filters for sargable-pushed conditions
to supplemental edges (which we handle as delayed predicates).

In particular, this fixes a bad misestimation in DBT-3 Q9 by about
a factor 10,000. Q9 still is underestimated, but that is due to
lack of proper use of statistics on multi-part indexes: If we have
an index on (a,b) and have a=C1 AND b=C2, we estimate those two
predicates' selectivity independently, instead of using the selectivity
information from the composite index.

Change-Id: Ib9cae24449680a5648bad84a4c0a0a3a622d198e